Update the release plugin to use the central publishing API directly (#3710)
Following the recommendation of the central publishing portal, this
updates our publishing plugin from Nexus to jReleaser. See:
https://central.sonatype.org/publish/publish-portal-gradle/

The way that this works is that now, the "publish" command will publish
artifacts locally to a staging repository that is in the build
directory. At the conclusion of the build, the `jreleaserDeploy` task
will take those artifacts and send them up to maven central. It will
sign them before it does so, which is why this also removes the manual
signing configuration that we'd previously had in our
`publishing.gradle` configuration.

I've tested this out by applying these changes to a parallel repo I had.
I was even able to validate that it could deploy to maven central, but I
had configured that other repo to only uplaod artifacts. This is
supposed to both upload and then publish the artifacts in the "deploy"
step. I have not been able to actually validate that that worked, for
obvious reasons. I can see locally that when I call `publish`, it
publishes to a staging folder in the build directory.

This resolves #3709.