Increase heap size for JReleaser (#3720)
This increases the memory limit to handle our artifacts after some
builds failed with "out of heap space".

While I was here, I also removed the `ReleaseNotes` updates for the
release that didn't actually publish, and adjusted the JReleaser
publishing directory to exclude the deploy directory, which includes the
artifacts.