Remove JReleaser publisher (#3721)
We weren't able to get the JReleaser to publish, even with higher heap
space. This PR reverts the work, and gets us back to where we were. In
that world, the release was also broken, failing with errors when it
would close the repo. We could, however, publish the change manually,
whereas this is currently failing before upload. So this may be better,
if we can't figure out how to make that other solution work.