Replace nexus publishing plugin with direct call to central sonatype uploading URL (#3723)
This replaces the usage of the nexus publishing plugin with direct calls
to the sonatype central publishing API. It makes the following
modifications:

1. Each published project will write to a central staging repo if we
have enabled central publishing.
1. A new task, `createMavenCentralBundle`, will go through and scoop up
all of the relevant files from the staging repo to produce the final
bundle
1. An additional task, `publishMavenCentralBundle`, will upload the
bundle to the central publishing API (see:
https://central.sonatype.com/api-doc and
https://central.sonatype.org/publish/publish-portal-api/#uploading-a-deployment-bundle)
1. The release action now calls the new `publishMavenCentralBundle` task

This should allow us to retire our usage of the legacy API. Locally, I
was able to validate that the bundle we get back from
`createMavenCentralBundle` has the same layout as one generated by the
JReleaser tool. Running the upload task, I was able to validate that I
get an invalid token error, which is as good as I could get without
giving it the proper credentials.

This resolves #3709