Remove full file load from publish step (#3724)
This updates the logic in the `publishMavenCentralBundle` task so that
it no longer loads the full contents of the file into memory. It does
this by maintaining a buffer and then writing the contents of the buffer
into the stream, which allows us to avoid materializing the full element
into memory. This is to try and fix a Java heap memory error that we saw
previously (see:
https://github.com/FoundationDB/fdb-record-layer/actions/runs/19108664941/job/54605649248).
    
While I was here, I also updated the boundary for the HTTP multi-part
form from the current time millis to a UUID, in response to a comment
from the previous PR
(https://github.com/FoundationDB/fdb-record-layer/pull/3723#discussion_r2495094985).