Disallow insertion of vectors of different precision and dimensions. (#3730)
It is possible to `INSERT` a vector with a different precision and
dimensions than the ones defined in the schema. This is fixed by doing
exact check of precision and dimensions in
`PromoteValue.isPromotionNeeded` which does, besides checking promotion
feasibility, type (in)compatibilities, which we may want to move to
somewhere else later on.

This fixes #3729.