Revert publishing directly via sonatype API (#3728)
This reverts #3723 and #3724 and returns us back to publishing via the
nexus publishing API. The impetus for this is that the last approach got
stuck in a publishing state after the upload (see:
https://github.com/FoundationDB/fdb-record-layer/actions/runs/19117047380/job/54633288001,
which succeeded, but for which we didn't get artifacts actually
published despite the server upload succeeding).

When we are in this mode, we expect the build to fail with a 400 error
letting us know that we didn't upload anything. However, if we didn't
`close` the staging repositories (which is the step that is failing),
then legacy API would not be expected to upload anything to the central
repo. So we need to continue to do that part even though we think it
will fail.