Revert Identifier Translation for Protobuf Compliant Names (#3726)
This reverts #3696 and #3706, which together were trying to handle
translating the SQL names to protobuf identifiers. This had some issues,
especially when it came to function and view name handling, and so we
want to redo this in a slightly different way. While we work on that
fix, this backs out the change.