Change DataInKeySpacePath to contain a KeySpacePath instead of ResolvedKeySpacePath (#3716)
It was unnecessary to have the ResolvedKeySpacePath, and we want to be
able to import into a different cluster, where the resolved values are
different from the exported data.
This means that DataInKeySpacePath has a few more fields, but makes it
more versatile going forward.
This also changes `exportAllData` to use `mapPipelined` to resolve the
path, so that `DataInKeySpacePath` does not contain a future.

This is in support of #3573