Make Lucene partitioner thread-safe (#3699)
Lucene partitioner is not thread safe when records are updated
concurrently.
The partition `count` and boundaries are not thread safe and can skew in
multi threaded updates.
The changes include protecting the partition metadata changes with a
`keyspace` lock.
The `LuceneIndexMaintenanceTest.concurrent*` tests have been extended to
run in partitioned setup and are now passing.
In addition, a bug where empty partitions were not removed was
uncovered, and a fix was introduced, with accompanying tests. Empty
partitions (first, middle and last) are removed with their index data,
ensuring at least one partition is always left behind.


Resolves #2990

---------

Co-authored-by: Scott Dugas <sdugas@apple.com>
Co-authored-by: Scott Dugas <scott.dugas@gmail.com>