Create test showcasing bug with enums and index scans
This adds a test case to showcase the problem with #3734. It modifies `enum.yamsql` with a new test case analogous to the one in the issue. There are two new queries that are planned with an index scan over an index on an enum valued field, and both of them fail with the same `NullPointerException` that is alluded to in the bug report.