Create test showcasing bug with enums and index scans (#3735)
This adds a test case to showcase the problem with
https://github.com/FoundationDB/fdb-record-layer/issues/3734. It
modifies `enum.yamsql` with a new test case analogous to the one in the
issue. There are two new queries that are planned with an index scan
over an index on an enum valued field, and both of them fail with the
same `NullPointerException` that is alluded to in the bug report.