Reintroduce type renaming with more resilient handling for unexpected protobuf names (#3770)
This reintroduces protobuf type and field renaming. The first version
was added in #3696 and #3706 and then taken out by #3726. A second
version was introduced with #3736 and then removed by #3767. This adds
back a variation based on the data from #3736 but updated to be more
resilient to unexpected names in existing meta-data objects.

The issue with #3736 before is that if a type existed in the meta-data
that was not correctly escaped (e.g., `Type__Blah` instead of
`Type__0Blah`), then it would fail to match the type during querying.
This was a problem even if the type wasn't actually involved in a query
because of how matching worked on the `FullUnorderedScanExpression`,
meaning that any query would fail to plan if any type in the meta-data
was so written.

This makes things more resilient. We now do a bit more work to associate
a type with its original name from the protobuf file if one is provided,
recording both the user-visible name and the original storage name. The
only places that we now generate new protobuf compliant names is when we
construct a `Type` object. In all other cases, we only go from the
storage name to user-visible names.

We still do rely on the fact that we can correctly predict the expected
user-visible name by running the de-escaping logic. At some point, we
may need to have a more complicated mapping, especially if we want to
support more arbitrary names. That is left as future work. I could also
see us wanting to do a bit more refactoring to better encapsulate this
transformation.

The new test modifications made to `valid-identifiers.yamsql` cover
those cases by adding new types with names that would not have been
generated by any DDL statement, and then validating that (1) those do
not disrupt correctly constructed queries and (2) that the problematic
types can themselves be queried.

In addition, this addresses some shortcomings with the match candidates
where `FieldKeyExpression`s (which use the internal names) would
sometimes be used to generate match candidates which referenced the
internal name directly. This fixes that by plugging those gaps. There
are additional queries in `valid-identifiers.yamsql` that are designed
to cover those matches.

---------

Co-authored-by: Youssef Hatem <y_hatem@apple.com>