Deprecate SQL type codes in JDBC (#3772)
With the `java_sql_types_code` field usages being removed in #3651 (in
favour of richer Type in JDBC), we are now in a position to deprecate
it. It also cleans redundant type infos that were present in `Parameter`
proto object. The latter had `java_sql_types_code` and new `type`
fields, and yet newer `ParameterMetadata` which itself has the
`java_sql_types_code` and `type` fields. To declutter, this PR deprecate
all except `type` in metadata. Note any of the type info is not being
used currently and hence it is safe to deprecate all in a single step

 This PR does the following changes: 

1. removes the setting of `java_sql_types_code` in ColumnMetadata,
Parameter and ParameterMetadata. Marks them as deprecated.
2. removes the setting of `type` in Parameter. Mark it as deprecated.