Remove recursive type name preservation when parsing meta-data objects
We previously had logic to preserve type names on all nested field types when we parse a file descriptor and turn it into a schema template. However, that can lead to problems if there are multiple nested types with the same name. This adds a yaml test with that configuration, and it also modifies the schema template parsing logic so that we no longer attempt to do this.

The type name preservation is still done on top level types, as that is still necessary in order to preserve index type names as well as to generate the FUSE expression off of the correct type names. However, most of the other locations that referenced the type name have been cleaned up so that they work on only the `Type` object.