YAML test framework now uses the packaged dependencies from the meta-data proto if included (#3793)
This updates the YAML test framework so that it can rely on the included
dependencies that are a part of the meta-data definition if there is
one. This is a test-only change, with the main change being to modify
the YAML test framework so that it doesn't attempt to get a dependency
from the classpath if it doesn't need to. The meta-data proto loader
will then handle building any of those included dependency files into
`FileDescriptor`s without the testing framework needing to know how to
do that.

This also adds a new YAML file to a special directory and a test that is
separate from the other `YamlIntegrationTests`. It is separated out
because the test isn't about running an integration test on our system,
but it is instead there just to test that we can operate on such a
meta-data JSON definition. The new test comes with a meta-data
definition which requires two `FileDescriptor`s, both of which are in
the meta-data proto.

This resolves #3792.