Handle multiple updates to the same record in a transaction (#3722)
The issue at hand is that when running multiple update operations in a
single transaction, the partition's document counts and the PK-segment
index may get into an inconsistent state. The root cause is that the
first update in the transaction clears the doc from the Lucene index and
the PK index. Since the changes are not flushed, the IndexWriter has
them cached in the NRT cache. The second record update would then not
find the record in the PK index (because the segment has changed but the
IndexReader does not yet reflect that) and therefore the delete is
skipped, including updating the partition count. Note that it does
attempt a delete-by-query that actually removes the doc from the Lucene
index, but since we can't know that, the partition is not updated.
The solution is to `refresh` the `DirectoryReader` when doing an update,
so that any previously written changes are showing up. The refresh
operation uses `DirectoryReader.openIfChanged` that is more efficient in
resources than using a brand new `open` call.

Resolve https://github.com/FoundationDB/fdb-record-layer/issues/3704