Introduce Serialization format for DataInKeySpacePath (#3747)
This adds a new `KeySpacePathSerializer` that supports serializing
`DataInKeySpacePath`.

In addition, I discovered that `KeySpacePath` doesn't actually validate
the values being added.
I thought about adding the validation to the add method, but felt that
was too risky, but I added the validation to the deserialization. This
means that if you are currently using `KeySpacePath` with incorrect
values, you may not be able to import the data. See: #3746


Resolves: #3745

---------

Co-authored-by: ohadzeliger <70664918+ohadzeliger@users.noreply.github.com>