Introduce PlannerEventListeners class for emitting planner events
This commit adds a new class and associated interface, PlannerEventListeners, which is going
to be used as the new entry point throughout the Cascades planner to emit planner events
to classes that are interested in processing these events, which implement the
PlannerEventListeners.Listener interface. The PlannerEventListeners use a thread-local list to keep
track of the different listener instances to dispatch events to.