IndexingMerger: lessen work and retry by default (#3732)
Catching an exception, the IndexingMerger attempts to detect if it
should lessen work and retry. Currently it is an include list of known
"retyable" errors. If the error is not detected correctly, or not
predicted, it will not retry.
Being a background process, however, it is better and safer to retry
after timeouts and after any FDB exception.

  Resolves #3731