Fix internal errors from version queries: Part 1
This begins the work necessary to fix internal errors related to `VersionValue` queries. The ultimate goal is to remove `VersionValue`, but for now, what this does is it modifies the way the various operators that sit on top of record fetches and has them copy out the version into a field, called `__ROW_VERSION`. Once that's done, then in a future version (say, 4.9), we'll be able to modify the `PlanGenerator` so that it references that field. Note that we'll need to wait for that rather than do it immediately for backwards compatibility reasons. If we took the final version right away, then any plan that wants the version when sent to an older version that doesn't do this copying will be unable to process that `FieldValue`.

This is in furtherance of #3796.