Set timeout on PRB jobs (#3810)
I saw a PRB take 4 hours due to an issue with a test. The default is 6
hours, but I think that is too long and would just be wasting resources
if a test has a deadlock.
I looked at a couple recent PRBs and doubled their times to put in
timeouts. Probably if we start to hit any of these timeouts we'll want
to parallelize our tests more.
I saw:
core-tests: 18m
lucene-tests: 15m
other-tests: 32m
style: 13m
coverage: 1m