Consider number of predicates at each level in rewriting cost model (#3681)
This PR improves the rewriting cost model by making it consider the
number of predicates in each expression at each level of the expression
graph, instead of comparing graphs based on the height of highest
expressions in the graph that contains any predicates.

The reasoning for this is to push the cost model to prefer expressions
with the same number of predicates, where some predicates were pushed
from higher levels in the QGM to lower levels, which can lead to
producing plans with more specific index key comparisons in the planning
phase.

Example: the rewriting cost model should consider the expression: 
```
SELECT sq1.a FROM (SELECT a, b FROM T WHERE a = 42) sq1,  (SELECT a FROM T2) sq2
```
to have a lower cost than:
```
SELECT a FROM (SELECT a, b, d FROM T) WHERE a = 42 AND EXISTS (SELECT a FROM T2)
```

The rewriting cost model is improved by considering two
`ExpressionProperty` when comparing two expressions: the existing
`NormalizedResidualPredicateProperty` property which calculates the
total number of conjuncts in the combined query predicate across the
entire expression tree at all levels and a new property
`PredicateCountByLevelProperty` which calculates the number of
predicates at each level of the expression graph. In addition,
considering the property `PredicateComplexityProperty` is no longer
necessary as what it checks for (the most complex predicate across the
entire expression tree) is already checked indirectly by
`NormalizedResidualPredicateProperty`.

## Performance impact

Some basic analysis of the performance impact of considering those
properties can be found in
https://github.com/FoundationDB/fdb-record-layer/pull/3681#discussion_r2632080223.